Exercise 10.8: the sign in form
Implement a sign in form to the SignIn component we added earlier in the SignIn.jsx file. The sign in form should include two text fields, one for the username and one for the password. There should also be a button for submitting the form. You don't need to implement a onSubmit callback function, it is enough that the form values are logged using console.log when the form is submitted.

Remember to utilize the FormikTextInput component we implemented earlier. You can use the secureTextEntry prop in the TextInput component to obscure the password input.